requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.7.0
boto3==1.34.56
botocore==1.34.56
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
Flask==3.0.2
Flask-Cors==4.0.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
inquirerpy==0.3.4
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
orjson==3.9.15
paramiko==3.4.0
pfzy==0.3.4
pinecone-client==3.1.0
prettytable==3.10.0
prompt-toolkit==3.0.43
py-cpuinfo==9.0.0
pycparser==2.21
pydantic==2.6.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tomli==2.0.1
tomlkit==0.12.4
tqdm==4.66.2
tqdm-loggable==0.2
typing_extensions==4.10.0
ujson==5.9.0
urllib3==2.0.7
uvicorn==0.27.1
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
